fix updating of toolbar combobox on zoom changes
let kword follow changes - zoom works again

svn path=/trunk/koffice/; revision=657184

+++ kword/part/KWView.h
@@ -100,7 +100,7 @@ private:
     void updateZoomControls();
 
 private slots:
-    void viewZoom( KoZoomMode::Mode mode, int zoom );
+    void viewZoom( KoZoomMode::Mode mode, double zoom );
     void updateZoom();
     void editFrameProperties();
